Add NAudio utils, formatting
