+++ NeuralCodecs.Core/NeuralCodecs.Core.csproj
@@ -6,7 +6,12 @@
     <Nullable>enable</Nullable>
 	  <IsPackable>false</IsPackable>
 	  <AssemblyName>NeuralCodecs.Core</AssemblyName>
+	  <Platforms>AnyCPU;x64</Platforms>
   </PropertyGroup>
+
+  <ItemGroup>
+    <Compile Remove="Utils\TensorUtils.cs" />
+  </ItemGroup>
 
 	<ItemGroup>
 		<AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleToAttribute">
@@ -15,7 +20,7 @@
 	</ItemGroup>
 
 	<ItemGroup>
-	  <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="9.0.4" />
+	  <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="9.0.5" />
 	  <PackageReference Include="NAudio" Version="2.2.1" />
 	</ItemGroup>
 </Project>
